requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
requests==2.31.0
apscheduler==3.10.4
python-chess>=1.9
